requirements.txt
google-api-python-client==1.6.6
httplib2==0.11.3
oauth2client==4.1.2
pkg-resources==0.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.2
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
